+++ include/mppic/critics/goal_critic.hpp
@@ -22,7 +22,7 @@ public:
 
 protected:
   unsigned int power_{0};
-  double weight_{0};
+  float weight_{0};
 };
 
 }  // namespace mppi::critics
